requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==4.2.13
django-ninja==1.1.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pydantic==2.7.1
pydantic_core==2.18.2
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
